requirements.txt
django==1.9.2
django-custom-user==0.5
djangorestframework==3.3.2
djangorestframework-jwt==1.7.2
django-filter==0.12.0
